requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.3.2
openpyxl==3.1.5

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# PDF processing
pypdf==3.17.0
PyMuPDF==1.23.8

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-documentai==2.20.0

# Fuzzy matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Logging and monitoring
tqdm==4.66.1